Fixing problem when restoring Profile Tables after server restart
Update Issue 2570
Status: Fixed
Applied a fix.

The issue was due to the new container lifecycle, deployment happens when SLEE is initialized, but not yet started. When the stored tables are being restored (during deployment), it was trying to create the activities, which failed since the SLEE was stopped.

Switched the flag not to create the activity, as it will be done at a callback informing that SLEE is running.

Please confirm it works for you.

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/jain-slee@18471 bf0df8d0-2c1f-0410-b170-bd30377b63dc