Fixing "Acct-Application-Id is duplicated in ACA"
Update Issue 2982
Status:Fixed
This issue was caused due to fetching Application-Id from header and creating the Acct-Application-Id AVP also.
It was wrong, as if the Application-Id was from a vendor, as the header has no vendor-id, it would create an Acct-Application-Id instead of the proper Vendor-Specific-Application-Id AVP.

Maybe as a side-effect to these fixes, the Application-Id header value does not match the one in AVPs. An issue has been created for that (Issue #3045).

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/jain-slee@21501 bf0df8d0-2c1f-0410-b170-bd30377b63dc